ajout fdf

+++ libft/ft_lstsize.c
@@ -6,13 +6,13 @@
 /*   By: jblancha <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/28 02:34:36 by jblancha          #+#    #+#             */
-/*   Updated: 2016/11/28 02:34:53 by jblancha         ###   ########.fr       */
+/*   Updated: 2016/12/14 18:55:42 by jblancha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_lstsize(t_list *lst)
+int		ft_lstsize(t_list *lst)
 {
 	size_t i;
